Moved database service functions from pkg/dbdiscauth to cmd/dbdiscauth, renamed functions to exported versions (GetInstancesFromDB, DeleteServiceFromDB, AddServiceToDB), updated AddServiceToDB to use BroadcastService instead of SendService, added json import, removed pkg/dbdiscauth import, and removed commented-out legacy code from redfishread.go
